requirements.txt
bcrypt==2.0.0
cffi==1.4.2
decorator==4.0.6
dnspython3==1.12.0
Flask==0.10.1
Flask-Bcrypt==0.7.1
Flask-Login==0.3.2
iniherit==0.3.4
ipython==4.0.2
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
path.py==8.1.2
pexpect==4.0.1
phonenumbers==7.2.2
pickleshare==0.5
ptyprocess==0.5
pycparser==2.14
pyIsEmail==1.3.1
python-slugify==1.1.4
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
Unidecode==0.4.18
Werkzeug==0.11.3
wheel==0.24.0
yoyo-migrations==5.0.1
